requirements.txt
# QuizSense AI - Required Packages
# Compatible with Python 3.11, 3.12, 3.13

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Azure OpenAI
openai>=1.12.0

# Database
aiosqlite>=0.19.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.6.0

# HTTP Requests
httpx>=0.26.0

# Date/Time
python-dateutil>=2.8.2

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3